feat: add advertisement banner for explore page with session tracking
